Allows to repair the MSI package even when the original installation package is no longer available.
Adds an additional source of the resiliency by attempting to create a symbolic link to the locally cached MSI package (%WINDIR%\Installer) or a hard link/full copy to the original installation MSI package in the specified directory.

WIXSHARP_RESILIENT_SOURCE_DIR property can be used to configure the target directory. INSTALLDIR property is used by default.

Windows 7 is shipped with the Windows Installer version 5.0, which unlike the previous versions of the windows installer caches the entire MSI, including internal CAB files. Unfortunately the complete cached MSI package is not used for repairs, a call is made to the original source which might not be available.

See also:

https://docs.microsoft.com/en-us/windows/desktop/msi/source-resiliency
https://www.symantec.com/connect/articles/reducing-windows-installer-disk-wastage-windows-7