An empty directory remains after the uninstall #464
Steps to reproduce:

Use a real certificate with the "Certificate" sample
Install/uninstall the MSI
Observe that an empty "%ProgramFiles%\My Company" remains after the uninstall
The issue is related to components without files or registry elements, in this particular case a component with the Certificate element. The WixSharp already contains a workaround code to inject "RemoveFolder" for a parent directory of the component, but it looks like it's not enough, it should be done for all chain of parent folders.