Reworks the Attributes dictionary and AttributesDefinition string to be more deterministic
This ensures that the dictionary is the source of truth and allows the string AttributesDefinition to be built on-the-fly. There is a minor performance implication here, but considering the usage scenario this seems reasonable to prefer correctness over speed.

The specific case this resolves (I'm sure there are others) is a situation where you might have:
```csharp
var entity = new File
{
    Attributes =
    {
        { "Checksum", true.ToYesNo() },
        { "Component:SharedDllRefCount", true.ToYesNo() }
    }
};
var initialCount = entity.Attributes.Count;
entity.ComponentId = "foo"; // Logically, this should add an attribute
Assert.Equal(initialCount + 1, entity.Attributes.Count); // Before the patch this fails.
```

Adds a new test theory that asserts deterministic behavior when using Attributes and AttributesDefinition interchangeably